requirements.txt
requests
sqlalchemy-1.4.0b2
numpy
pandas
pandas_market_calendars
arrow
tzlocal
demjson
plotly
dash
dash-bootstrap-components
dash_daq
tqdm
jqdatapy
baostock
yfinance